backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Testing
pytest==7.4.4
httpx==0.26.0

# MCP SDK (install with: pip install @modelcontextprotocol/sdk)
# Note: This package may need to be installed separately as it's not available in standard PyPI
